fix: TLS certs command line parameter handling
Previously, the TLS certs command line parameter and env var equivalent had different formats: the command line flags had to be repeated, while the env var used comma separated values. This was confusing since the documentation simply says that it's supposed to be in CSV format, but if you tried that with the command line flag then it wouldn't work as expected. Using a new method I added to edge-core, now both method work the same way.

Also used this new method to simplify the logic for the auth-type flag/env var.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>